Tracker: Activate window on pose mouse up
... if we are not in focus follows mouse mode and not dragging.

This copies the BeOS R5 behavior of activating the window on
mouse down if you click on blank and mouse up if you click on
a pose. This allows you to start a drag from a background window
by clicking on a pose and dragging it without activating the
window. If you wait too long to start a drag before moving your
mouse after pushing down the mouse button (double click speed)
you'll trigger a long press which will activate the window.

Define kDragSlop in Tracker.h and use that value (3) throughout
Tracker, including in PoseView, to determine the number of pixels
the mouse must move to be in a drag.

Fixes #8824

Change-Id: If3c82d3f0acc4603dc6a8037d399975708436350