Update patch set 3
Patch Set 3:

> As a side note, I see some var == 0.0 on adjacent lines. Isn't that a bad thing. My understanding is that you want a margin of error on floating point?

The agg code is known to have a lot of such warnings on PVS and Coverity, indeed. Comparing to 0.0 exactly can be on purpose, if one is sure that the value was explicitly set to 0 and not the result of some computation.

So it takes time to review all the code and decide if this is intentionally comparing to explicit 0, or if we should add some margin.

Patch-set: 3