Update patch set 1
Patch Set 1:

> Why do we need to manually Activate() here? Shouldn't the OS do that on its own?

My understanding is that the reason we don't use the default window activation in Tracker here is because we don't want to activate the background window on mouse down while dragging because that would potentially cover up the currently active window you are dragging to. The current code annoyingly only activates the window if you clicked on the blank area and not on a pose to preserve this drag and drop behavior.

On #8824 there were a couple of solutions proposed, one was to activate the window on MouseUp instead of MouseDown. I implemented this solution and it does indeed fix the problem, but, it introduces a delay that humdinger would object to. I installed an unadulterated BeOS R5 Pro into a VM (black and white only) to make sure I wasn't copying some Max Edition v4 change and Tracker does indeed activate a window on pose mouse down.

Another solution was to add a key-combo to activate the window but I reject that solution as too complicated.

I went into Open Tracker code and checked what it did and I essentially copied what OpenTracker was doing which was to introduce a delay on mouse down to check for a drag before activating the window.

https://github.com/HaikuArchives/OpenTracker/blob/master/opentracker/tracker/PoseView.cpp#L6200

Patch-set: 1
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
CC: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>