Fix qt mark on TR-F layout/Add TR Lira Symbol/Clear whitespace
On Turkish F layout Shift+2 produces single quote, although it
should return double quotation mark. Probably this is a result
of an error whilst copying from Turkish Q layout.

Turkish Lira Symbol has been added to the both layouts,
produceable with ALT GR+T, as this is the standard combination
on Turkish layouts.

I've also noticed that most of the keys produce a whitespace with
the ALT+GR modifier, fixed that one as well.