Update patch set 1
Patch Set 1: Code-Review+1

> I've also noticed that most of the keys produce a whitespace with the ALT+GR modifier, fixed that one as well.

This is my fault, I changed it so that if you don't have anything mapped to a map (in this case the option map) it will fall through to the normal map. Some of the older keymaps put a space in the normal map for some reason so those need to be be taken out for this feature to work properly.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1