Update patch set 2
Patch Set 2: Code-Review+1

I believe that I've come up with a fair compromise between not activating background windows on a drag and activating on mouse down instead of mouse up in patch #2. If you click on a blank area the window will activate immediately as before, but if you click on a pose the window will activate after 1/4s unless you start a drag by moving the mouse 3px or more. 1/4s seems to be a comfortable amount time to start a drag without introducing too long of a delay and 3px is used elsewhere in Tracker as the drag threshold.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1