Update patch set 4
Patch Set 4: -Code-Review

> Patch Set 4: Code-Review-1
> 
> I take back me general approval of my former comment. :)

On Patch Set 1 I wrote

>> Patch Set 1
>> On #8824 there were a couple of solutions proposed, one was to activate the window on MouseUp instead of MouseDown. I implemented this solution and it does indeed fix the problem, but, it introduces a delay that humdinger would object to.

I believe that there is a tiny yet perceptable delay that you are objecting to between when you release your mouse button and MouseUp executes. On the previous patch that you gave approval to even though the window was activated when you lifted your mouse button it still technically happened in MouseDown. In Patch Set 4 the window activates not only when you release your mouse but in the MouseUp method as well. I'm not sure how we can elminate this message delay without blocking, and this is the delay I was referring to in Patch Set 1.

> After playing a bit more with the latest patch, it does feel weird. Any other app's window pops to the front instantly on mouse down. Only Tracker windows have the small delay as they pop up on mouse up. Do switch for some time between Tracker and other apps' windows and you feel that something is off...

If it's any consolation to you, this is exactly how macOS works. Pretty much all apps activate on mouse down including Finder, unless you click inside the pose view or whatever they call it and then the window activates on mouse up for exactly this reason.

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0