Update patch set 4
Patch Set 4:

I can imagine it being frustrating after putting in quite some effort, but TBH if it avoids inconsistencies and delays when compared to switching between other apps' windows, I'd rather have Tracker behaviour stay as it is.

I can even see it as a feature: Click anywhere in the white area of a Tracker window to pop it to the front, click a file icon/name and the window stays in the back. That way you can open a file or rename it without losing focus of the currently active window.

Patch-set: 4
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>