Terminal: Fix double-freed memory
The base problem was, that every time I opened KeyMap while a Terminal
was opened too, it eventually crashed. I followed along the trace and
found that the code first frees the keymap and keymapchars in the
_UpdateKeymap() and then calls SetKeymap for every tab opened which
also tries to free the old keymap and keymapchars.

TermWindow's _UpdateKeymap() method first frees the old keymap and keymapchars:

delete fKeymap;
delete[] fKeymapChars;

and then calls SetKeyMap for every tab which tries to free the old keymap and keymapchars again:

for (int32 i = 0; i < fTabView->CountTabs(); i++) {
	TermView* view = _TermViewAt(i);
	view->SetKeymap(fKeymap, fKeymapChars);
}

So TermView simply is not responsible for the memory, it merily has a pointer.