Update patch set 1
Description set to "The base problem was, that every time I opened KeyMap while a Terminal was opened too, it eventually crashed. I followed along the trace and found it that the code first frees the keymap and keymapchars in the _UpdateKeymap() and then calls SetKeymap for every tab opened which also tries to free the old keymap and keymapchars."

Patch-set: 1
Patch-set-description: The base problem was, that every time I opened KeyMap while a Terminal was opened too, it eventually crashed. I followed along the trace and found it that the code first frees the keymap and keymapchars in the _UpdateKeymap() and then calls SetKeymap for every tab opened which also tries to free the old keymap and keymapchars.
Tag: autogenerated:gerrit:setPsDescription