Update patch set 1
Patch Set 1:

> Did this test help at all? Should I add this comment to #7434?

Yes, that would mean this particular issue is solved.

It remains that translating application name is a great things in some locales, and culturally unfit in others, so I'm still not sure what to do here. Maybe we should survey users to see if there is a clear choice for each locale or if we really need to keep this as an option.

What worries me about it being an option is that it complicates documentation efforts (now each application has two names depending on user settings) so I would prefer that this would be set once and for all for each locale. I'm not sure if we can agree on a fixed setting for all of them, however.

Patch-set: 1