Update patch set 1
Patch Set 1: Code-Review-2

So, after reviewing more closely, I agree with korli that this does not seem to be the right fix.

The packets sent to an USB-MIDI device get an extra header including a "virtual cable" number, allowing multiple streams of events to be sent to the same device. After this header, there are always 3 bytes of data for each event, even if the actual MIDI data is only one or two bytes (padding is added).

So, with this patch, we are now replacing the missing bytes for the last packet with 0s. This is not what is expected. If the driver really needs to handle truncated packets, it will have to buffer them, wait for an upcoming write (on the same file descriptor) to complete the packet, and only then, send the data. Until this is implemented, it seems better to not send the packet at all and return an error instead.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2