support/String: Revert 6c67c7d63 to fix double-free
6c67c7d63 was attempting to fix a leak caught by a static analysis
tool, but it actually just introduced a double-free bug. Running
`UnitTester BString` will result in a crash..

The original code was correct because, in the event that realloc()
fails in BString::_Resize(), the value of fPrivateData is still
retained. It will be freed by the destructor of BString only if
fPrivateData is not shared by another BString instance, since BStrings
are copy-on-write.

Note that while the change in 6c67c7d63 caused tests to fail, that
doesn't mean those tests are ideal. They only trigger
BString::_Resize() to fail because they depend on implementation
details of hoard2 where on implementation details of hoard2 which
limit allocations via malloc() to 1GB. Most malloc() implementations
will allow allocations of arbitrary sizes using anonymous
mappings (mmap on Linux, or create_area() in Haiku). This is a much
bigger change, so for now I'm just adding some comments so that we can
revisit these tests if we make a change to the allocator.