Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> free() + malloc() would be simpler and faster.

It would be more sophisticated, but in an ideal case, realloc would be faster since it would only change the header of that block. Though in other cases it is much slower because of the copying procedure of realloc.

Patch-set: 2
Reviewer: Gerrit User 1000170 <1000170@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>