Update patch set 1
Patch Set 1:

Does it at least fix that bug with atan2?

There is clearly no support for m68k, how would we go about adding that?

I find it surprising that there does not seem to be common code for specific long double formats, in particular, can you document a bit on how the code is structured for that? Here is how it works in glibc (not our import of it, which was not done correctly but that could be fixed):

- Each platform has mostly asm files to leverage specific support from the CPU
- There are few truly "generic" functions
- A lot of the code is sorted out by representation format, typically, you have implementations for 128bit long double (used by sparc and risc-v), 80bit long double (x86 and as far as I can tell, m68k), 64bit long double (arm and possibly arm64), etc.

In glibc these are just in different directories which the Jamfiles for each platform can reference, reducing code duplication. How is this handled in musl? Is the code duplicated for each architecture, or are multiple formats handled in the same "generic" file somehow?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>