Update patch set 2
Patch Set 2:

> I did not say that.

That = "how convoluted the headers are"?

OK. kallisti5 definitely did, anyway; I must have misremembered.

> This is likely to break ABI, unless proven otherwise you get my -2.

It is not likely to, and furthermore, it doesn't:
 * The symbols report is listed above, you can see for yourself there.
 * The musl files use our math.h, and in full-generics mode, GCC (8) issues no warnings about mismatched prototypes (a few functions not in our math.h issue warnigns about missing prototypes, and you can see these new symbols added, indeed.)
 * Various applications both GCC2 and GCC8 that use math functions still work just fine.

The only possible breakage here is the removal of weak_aliases on some of the functions, but this is now more in line with the specs on what functions should be weak and what not. Or if some applications actually depend on broken behavior.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>