Update patch set 4
Patch Set 4:

The docs is lost, but it looks like the output thread is processing all the inputs and outputs? In this case stopping it on disconnect wouldn't make sense.

Also, it looks like there's an AutoLocker belonging to the buffers, used by the output thread. We might need to add it to the disconnect handler instead (otherwise we might be clearing a pointer and writing to it simultaneously)

Now let's see if I can set up a dev environment...

Patch-set: 4
CC: Gerrit User 1000179 <1000179@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>