Update patch set 4
Patch Set 4:

> Moreover, "setlocale -c az -v Cyrl -s AZ" works.

Maybe I'm confused, but isn't "Cyrl" the script and should be set with -s ?

> "setlocale -c de" does not work because "de" is not a language.

You mean "country". You may be right, there is a (null) in the "country" column of --details.
I have to say, I'm a bit confused when some columns are blank: I never know which column it is of the 6...
code <TAB> language <TAB> country <TAB> script <TAB> variant <TAB> name

Maybe put a "-" in the blanks?

Patch-set: 4