Update patch set 3
Patch Set 3:

Nothing in the API says there is a particular z-order for sibling views. IIRC this was already discussed a few years ago and we were not convinced it was right to set one.

I may change my mind if this is shown to help with BeOS compatibility in some way.

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>