Update patch set 5
Patch Set 5:

> Patch Set 1:
> 
> Depends on https://review.haiku-os.org/c/haiku/+/2091.

FWIW, it would be better if you worked in branches, and did not keep all of your patches on one branch; as doing so means that every time you push, the change-requests are needlessly updated and it spams emails about it.

If you keep unrelated changes on separate branches, same as sending GitHub PRs, and push those, then they will of course have separate histories and will appear here as disconnected.

Patch-set: 5
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>