Update patch set 14
Patch Set 14:

I'm really sorry that I haven't found the time to reply earlier. I tried to remember why the removed code was there, but I can't (precisely). Before this change is commited, I would love for the commit message to be a lot more informative. I would hope that it was understood why the removed code was there in the first place, why it is wrong and in what situation it caused flickering. The app_server can have multiple drawing backends, some of which may be double buffered and some not. So perhaps the removed code is actually correct for some of these backends and instead of removing it, it should check some property of the current backend in the if-condition. I would guess that for single buffered backends, it reduces the delay between the expose event and background regions being cleared: Think of a window filled with a gray background view, which is drawn by app_server. If you enlarge the window, the time until the view is filled with its background color would be increased, since the window frame is rendered immediately. At least I am sure the removed code wasn't there by accident, and the commit message right now doesn't really give me confidence that it was investigated deeply enough.

Patch-set: 14