Update patch set 1
Patch Set 1:

> If I understand correctly, this is just a warning during build? If so, I'd prefer we keep it as a reminder to fix the problem.
> I have no strong feelings about this, as you can see I didn't give it a -1.

Hi.  This change will modify the build so that it takes this problem (treating object pointers as raw memory) to be a warning rather than error that would otherwise fail the build.  The warning will still be there and there is a ticket to perform the work to deal with the warning.  Is that OK to merge?

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>