Update patch set 16
Patch Set 16:

Thanks waddlesplash! So the reason for the code is indeed that expose events shall clear the background as soon as possible so that outdated stuff is removed from the screen as quickly as possible. What I am wondering is this: If the client set the view color to transparent, this should also prevent the background from being drawn by app_server even in this code-path. For views where app_server draws the background (non-transparent view color), if memory serves, the code does the following: If the client requested a redraw, than the background is cleared by app_server, *but* with a double buffered drawing engine setup, the back-to-front update is delayed until the client has drawn on top of the background that app_server has cleard. Whereas if the redraw as caused by an expose event (i.e. was not requested by the client) then the back-to-front update happens immediately. And that's what causes the flickering. The question is what looks better: Tearing due to out-dated stuff staying on screen longer (by removing this code), or no tearing and flickering instead. At least explain this in the commit message that another trade-off is now preferred.

Patch-set: 16