Update patch set 16
Patch Set 16:

Thanks X547 for investigating! My complaint is not with the change, I am fine with it and understand that this is a better trade-off. But please just copy&paste your explanation into the commit message. If someone looks at this in three years, it needs to be clear what you found out and why you've made this change.

Patch-set: 16