Update patch set 2
Patch Set 2:

As far as I remember, the idea is that on the app_server side, the views may already have other sizes. But the notification for the size change may arrive after the notification for the redraw request. So the Window checks the sizes and updates them in both events in order to avoid needless overhead. Can you elaborate on what problem you identified with this approach?

Patch-set: 2
CC: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>