Update patch set 2
Patch Set 2:

Hm, maybe I understand. But from what I remember, the update session is started from the client side *after* it received the update request from the server (and there is only one request for all views, handled by the window). So why would it be incorrect to update the view sizes before starting the update session? Maybe the problem is that the server has an even newer view layout then what is contained in the update request to the client? Then removing this code should make the problem worse.

Patch-set: 2