Update patch set 2
Patch Set 2:

In any case, consider this: The window is enlarged. App_server sends an update request to the client. The window redraws every view, but with the old size. Then the resize event is received, causing the window to re-layout and that in turn will trigger a whole bunch of update requests from the client. It would be much more efficient, if this is all handled with one single event.

Patch-set: 2