Update patch set 1
Patch Set 1:

Hi X547, I don't understand why this is necessary. This is really a long time ago and I have a hard time remembering, but I thought the idea was that once you do a "complicated" CopyRegion operation in the backbuffer, you can then just copy the changed area's bounding box to the front as it is.

Patch-set: 1
CC: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>