Update patch set 1
Patch Set 1: Code-Review-2

Long-double changes are OK.

However, the addition of the generics is not. As I already explained on IRC, these are supposed to be the builtins now. You sent me a paste of a gcc -E output that shows that the non-builtins were selected; which means that, somehow, __GNUC__ < 7, so your compiler is just very broken. That needs to be fixed.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2