Terminal: Fix crash for keystroke with unexpected modifiers
This patch fixes a segfault in Terminal that happens when you issue a
keystroke with unexpected modifiers. An easy way to repro this is:

 1. Set command key to super
 2. Set option key to meta in Terminal
 3. Issue `ctrl+alt+shift+<key>`

In DefaultState::KeyDown(), the branch that handles option key as meta
detects that additional modifiers were used along with the meta key
and tries to look up the keymap associated with that combination of
modifier keys.

This lookup in TermView::fKeymapTableForModifierse can return NULL,
since TermView::SetKeymap() sets a pre-defined set of (modifier combo)
=> keymap entries.

This patch just falls through if tehre is no keympa associated with
the provided combination of modifier keys, The very next check will
see numBytes <= 0 and will return from KeyDown, effectively swallowing
the unhandled keystroke.