Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+1
> 
> Looks good. I expect you've tested all those settings (send & receive).
> Tell us when you've got it all finished.
> Also, when you're ready, put a bullet list of the added providers in the commit message.

I don't possess accounts on all of the providers (primarily the ones that are being hosted by universities and educational institutions, whose services are targeted towards the faculty, as well as the students that are currently enrolled), so I am not in the position to send and receive e-mails for the most part.

What I did for the vast majority of the domains, however, was use official documentation provided by the providers and verify that the services indeed exist and were accessible. I also asked acquaintances whether they'd like to see their own provider added here and I did receive help regarding some of the settings. For the time being, I believe that I am finished, although I may add more providers in the future.

I can provide the links to the documents that I used, test the login function (with false credentials, just to figure out whether there's really a backend that actively rejects my login requests, although I think that this may be too extreme of a scenario, if you take the fact that these services are used daily and outdated information would be counterproductive or problematic for the organizations) or remove changes upon request.

Patch-set: 7
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>