Update patch set 9
Patch Set 9:

> Worst case: the config doesn't work and people have to tweak the mail_daemon settings manually (and hopefully report this so we can either fix or remove the ProviderInfo.

I also added most of the links with the help of a bunch of helpful people, who are actual users of the instances that use the included configurations.

However, I haven't been able to confirm namely the faculty configurations of TU Mannheim, the postserver instance of TU Dortmund, as well as TU Delft's configurations. But I did verify that all corresponding ports and services are live, including for the services that I just mentioned. I think that not adding the less accessible servers would be like adding half of the functionality.

I could try to look into it further by contacting the relevant departments and administrators that should be able to verify that the servers indeed function, which is really no trouble for me. However, I included those changes with reasons as to why they shouldn't just work in theory.

Patch-set: 9
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>