Update patch set 3
Patch Set 3:

> > That would be at the cost of losing our fast remote app_server implementation, however. The server-side drawing allows us to intercept all drawing commands and transfer them to another machine very easily and in a transparent way for applications.
> 
> OK, I didn't have that in mind. But what is a common use-case for the remote-app-server? I have never needed to use it, so I am really ignorant.

Anything where you need to connect to a remote machine and do things there. And it even works with an html5 client now, so you can connect to your Haiku machine from any web browser, how nice is that?

> 
> > It also avoids sharing the buffer memory between apps and app_server, which opens for other optimizations (no need to create a separate area for each application for security reasons) and keeps the way open for future hardware accelerated drawing in easier ways, I think?
> 
> But every BBitmap is already a shared area between app_server and an application. So how can security get any worse than before?

In the current situation, it's shared only for BBitmap. Applications don't normally get access to what's on the screen. Being able to peek at other applications or change what they display is an attack vector. This is why Windows will (annoyingly) blink the screen black before showing their security dialog for operations needing administrator privileges, for example.

So, if we allow each application to modify its window buffer directly, not only we need separate buffers for each application, but we also need each of these to live in a separate area (possibly shared with the BBitmaps for the same application, assuming we don't plan to implement a way to handover a bitmap to another app) so we can make sure an application cannot easily access another's window (and either draw on it or spy on what's there).

Currently this is solved for on-screen windows because the application does not get access to the buffer at all, so it's entirely owned by app_server.

Likewise for performance, if you do client-side drawing, this memory is shared so you need some IPC to synchronize access to it. I think this is where te "more CPU" impression comes from, because there is a lot of interlocking between the application and the drawing thread if you do this. If howeverm you keep app_server in control of both the drawing and flushing to the display, you can get easier synchronization, and more importantly you keep this asynchronous from the application, so the event loop there continues to run without being blocked.

Patch-set: 3