Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > > That would be at the cost of losing our fast remote app_server implementation, however. The server-side drawing allows us to intercept all drawing commands and transfer them to another machine very easily and in a transparent way for applications.
> > 
> > OK, I didn't have that in mind. But what is a common use-case for the remote-app-server? I have never needed to use it, so I am really ignorant.
> 
> Anything where you need to connect to a remote machine and do things there. And it even works with an html5 client now, so you can connect to your Haiku machine from any web browser, how nice is that?

Can you give me a concrete example? Whenever I needed to do something on a remote machine, I have used a CLI. For everything else, I have always sat in front of the machine. I am really lacking the experience and possibly the phantasy to imagine a use-case.

> > But every BBitmap is already a shared area between app_server and an application. So how can security get any worse than before?
> 
> In the current situation, it's shared only for BBitmap. Applications don't normally get access to what's on the screen.

Why would they need to get access to the screen? They need access to their Window-buffer, which could work exactly as BBitmaps work now. In fact, BDirectWindows have access to the screen right now, and they can indeed draw over everything.

> Being able to peek at other applications or change what they display is an attack vector.

BDirectWindows could do that right now.

> This is why Windows will (annoyingly) blink the screen black before showing their security dialog for operations needing administrator privileges, for example.

I am not sure that is true. I am 100% sure I saw a UAC prompt on my Windows laptop recently which did not darken the screen. I assume they do it for extra attention, at least recently.

> So, if we allow each application to modify its window buffer directly, not only we need separate buffers for each application, but we also need each of these to live in a separate area (possibly shared with the BBitmaps for the same application, assuming we don't plan to implement a way to handover a bitmap to another app) so we can make sure an application cannot easily access another's window (and either draw on it or spy on what's there).

But how is this any worse than BBitmaps right now?

> Currently this is solved for on-screen windows because the application does not get access to the buffer at all, so it's entirely owned by app_server.

Why would the application need access to the screen in BDirectWindow fashion for implementing buffered windows?

> Likewise for performance, if you do client-side drawing, this memory is shared so you need some IPC to synchronize access to it. I think this is where te "more CPU" impression comes from, because there is a lot of interlocking between the application and the drawing thread if you do this. If howeverm you keep app_server in control of both the drawing and flushing to the display, you can get easier synchronization, and more importantly you keep this asynchronous from the application, so the event loop there continues to run without being blocked.

This doesn't convince me at all. Right now we have the concept of the update session: The app_server requests a window to redraw, it starts the session which puts app_server in a special mode to try and avoid blitting stuff to the front buffer, all views in a client window render, the update session is finished and app_server copies back to front buffer for the region of the window. In the buffered scenario, it would work exactly the same, except app_server can use the old window buffer for any drawing it might have to do, allowing smooth and flickerfree updates. And once the window is done, a pointer in app_server is flipped to the new buffer, and the back to front copy can happen. No extra synchronization needs to happen, and there is a lot less communication to app_server if the client draws itself.

Patch-set: 3