Move source and development packages out of /system/packages on release images
Activating all these packages on live boot (especially live DVD where we
also have the write overlay eating up RAM) significantly increases
memory requirements for running Haiku, to the point that Installer may
fail to run. Move them outside the packages/ directory so they are not
activated.

Installer already skips these two directories when installing, so the
packages will not be copied to the target install. This is fine for
source packages, for development packages there should be a checkbox
allowing to enable/disable them in Installer. Currently the logic for
that is based on FS attributes which are never set by our buildsystem,
the next step is to review this and either add the attributes, or change
how Installer works to show the checkboxes.

Fixes #15621.

Change-Id: Ibfac14c43dbe708316a0d82ca1c9d49b5cba2aef