Update patch set 6
Patch Set 6:

Hello there; This has been in review for ~1.5 months.  I really appreciate the help so far with the review, but I would like to merge this and keep the momentum going on completing this larger piece of work.  I will merge this in a day or so unless there are any objections.

Patch-set: 6
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>