Installer: use Walter as a placeholder logo
For non-official builds, Installer had no logo. This led to a window
sizing bug which is not seen in the official builds (both nightlies and
releases) as they both use the Haiku logo there.

Use Walter as a placeholder logo in that case. The png is generated from
the wonderbrush file in the separate artwork repo.