usb_disk: handle large drives
The read_capacity_10 command returns the number of sectors as a 32bit
value. If that's not large enough, we need to retry using the
read_capacity_16 command instead.

We should probably review all other commands (read, write, ...) to see if
similar changes are needed. At least make really sure that we check for
overflows and don't end up writing the wrong sector.

Change-Id: Ic396eae2c2b65e525125d4e0a257694a4a13f576