Actually handle more than 3 mouse buttons
The code in input_server was pretty much all set for this, but there was
no way to configure the extra buttons. Add them to the mouse view in
Input preferences (up to 5 buttons are handled now)

Define the two new buttons to be B_FORWARD_MOUSE_BUTTON and
B_BACK_MOUSE_BUTTON as that's what they are usually used for.

Change-Id: I9091082277937d89b08464ff474e7bbb5db82401