Update patch set 1
Patch Set 1:

Hi stippi, do you know/remember what this device is meant for? Should we remove it, or are there plans for using it, in which case we should maybe tell input_server to ignore it because it's not an input device?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>