Update patch set 1
Patch Set 1:

Well we have to decide on a way to manage the lifecycle of SharedBufferList. Either it's the use of Get/Put and reference counting, and the class is deleted when the last reference to it is Put(). Or, we decide that the MediaRoster is in charge of deleting it through Invalidate when it goes away.

But, we can't have both. Otherwise, any thread that thinks "it's okay, I can use the buffer list because I have a valid reference to it" is wrong. At any time, the media roster can delete the class regardless of wether there are any references to it.

I don't know which of the two options makes more sense, but we have to decide on only one here.

So, yes, manual reference counting comes with the risk that someone forgets to release a reference and there is a leak. But on the other hand, it allows the ownership of the list to  be shared. The other option is to have one thread/class responsible for deleting it, but in that case we have to make sure it's done after other threads are finished, and right now that's not the case. If we go that way, then I don't see what the reference counting in SharedBufferList is for?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>