Update patch set 1
Patch Set 1:

It is hard to do a poll about this because if we ask the question in english, results will be biased towards people who know at least a little english. I think what would make sense is a poll on each of the -i18n language lists (and relayed to local user communities if they exist and the people on the list decide it should be enabled), and we should make this setting per-locale then.

Patch-set: 1